+++ R/aaaCranNlmixrThreads.R
@@ -25,7 +25,6 @@ aaaCranNlmixrThreads <- function() {
     data.table::setDTthreads(2L)
     Sys.setenv(OMP_NUM_THREADS = "2")
     Sys.setenv(MKL_NUM_THREADS = "2")
-    #rxode2::rxUnloadAll(FALSE)  # don't unload any models (seems to affect ASAN checks)
   }
 
 }